% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styler.R
\docType{methods}
\name{APAStyler}
\alias{APAStyler}
\alias{APAStyler,SEMSummary-method}
\alias{APAStyler,lm-method}
\alias{APAStyler,mira-method}
\title{A generic function for pretty printing in (semi) APA Style}
\usage{
APAStyler(object, ...)

\S4method{APAStyler}{lm}(object, digits = 2, pdigits, file, ...)

\S4method{APAStyler}{mira}(object, lmobject, digits = 2, pdigits, file, ...)

\S4method{APAStyler}{SEMSummary}(object, digits = 2, type = c("cov", "cor",
  "both"), stars = FALSE, file = ifelse(.Platform$OS.type == "windows",
  "clipboard", FALSE), sep = "\\t", ...)
}
\arguments{
\item{object}{An object with a class matching one of the methods}

\item{digits}{The number of digits to round results to. Defaults to 2.}

\item{pdigits}{The number of digits to use for p values. Defaults to digits + 1 if missing.}

\item{file}{An optional argument indicating whether the output should be written to a file.}

\item{lmobject}{an lm object the degrees of freedom of which can be used for conservative F tests}

\item{type}{A character vector giving what to print. Defaults to \sQuote{cov},
the covariances. Other options are \sQuote{cor} and \sQuote{both}.}

\item{stars}{A logical value whether to include significance values as
stars (*** p < .001, ** p < .01, * p < .05).}

\item{sep}{Character what the separator for the table should be. Defaults to tabs.}

\item{object}{\code{lm} object}

\item{object}{\code{mira} object}

\item{digits}{The number of digits to round results to. Defaults to 2.}

\item{pdigits}{The number of digits to use for p values. Defaults to digits + 1 if missing.}

\item{file}{An optional argument indicating whether the output should be written to a file.}

\item{object}{\code{SEMSummary} object}

\item{digits}{The number of digits to round results to. Defaults to 2.}

\item{file}{An optional argument indicating whether the output should be written to a file.}
}
\description{
A generic function for pretty printing in (semi) APA Style

A generic function for pretty printing in (semi) APA Style

A generic function for pretty printing in (semi) APA Style

A generic function for pretty printing in (semi) APA Style
}
\examples{
m <- SEMSummary(~., data = mtcars)
APAStyler(m, type = "cor", stars = FALSE, file = FALSE)
APAStyler(m, type = "cov", stars = FALSE, file = FALSE)
APAStyler(m, type = "both", stars = FALSE, file = FALSE)
APAStyler(m, type = "cor", stars = TRUE, file = FALSE)
APAStyler(m, type = "cov", stars = TRUE, file = FALSE)
APAStyler(m, type = "both", stars = TRUE, file = FALSE)
}

