% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styler.R
\name{APAStyler}
\alias{APAStyler}
\title{A generic function for pretty printing in (semi) APA Style}
\usage{
APAStyler(object, ...)
}
\arguments{
\item{object}{An object with a class matching one of the methods}

\item{...}{Additional argiuments passed on to methods.}
}
\description{
A generic function for pretty printing in (semi) APA Style
}
