% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{cramerV}
\alias{cramerV}
\title{Calculate Phi or Cramer's V effect size}
\usage{
cramerV(x)
}
\arguments{
\item{x}{A frequency table, such as from \code{xtabs()}.}
}
\value{
A numeric value with Phi for 2 x 2 tables or Cramer's V
  for tables larger than 2 x 2.
}
\description{
Simple function to calculate effect sizes for frequency tables.
}
\examples{
cramerV(xtabs(~ am + vs, data = mtcars))
cramerV(xtabs(~ cyl + vs, data = mtcars))
cramerV(xtabs(~ cyl + am, data = mtcars))
}
