% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{diffCircular}
\alias{diffCircular}
\title{Calculate the Circular Difference}
\usage{
diffCircular(x, y, max)
}
\arguments{
\item{x}{Numeric or integer values}

\item{y}{Numeric or integer values}

\item{max}{the theoretical maximum (e.g., if degrees, 360; if hours, 24; etc.).}
}
\value{
A value with the circular difference. This will always be positive if defined.
}
\description{
Calculate the Circular Difference
}
\examples{
diffCircular(330, 30, max = 360)
diffCircular(22, 1, max = 24)
diffCircular(c(22, 23, 21, 22), c(1, 1, 23, 14), max = 24)
}
