% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{.allmissing}
\alias{.allmissing}
\title{Determine which if any variables are all missing in a dataset}
\usage{
.allmissing(data)
}
\arguments{
\item{data}{A dataset to check each variable if all missing / non finite / zero character vectors}
}
\value{
\code{FALSE} if no variable(s) all missing, else an informative string message.
}
\description{
Internal function.
}
\examples{
JWileymisc:::.allmissing(mtcars)
cat(JWileymisc:::.allmissing(data.frame(a = NA, b = 1)), fill = TRUE)
}
\keyword{internal}
