% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styler.R
\name{.fround}
\alias{.fround}
\title{Function to round and format a number}
\usage{
.fround(x, digits)
}
\arguments{
\item{x}{the data to round and format}

\item{digits}{the number of digits to used}
}
\value{
a character vector
}
\description{
Function to round and format a number
}
