% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{egltable}
\alias{egltable}
\title{EGL Table function makes nice tables}
\usage{
egltable(vars, g, data, strict = FALSE)
}
\arguments{
\item{vars}{Either an index (numeric or character) of
variables to access from the \code{data} argument,
or the data to be described itself.}

\item{data}{optional argument of the dataset containing
the variables to be described.}

\item{strict}{Logical, whether to strictly follow the
type of each variable, or to assume categorical if
the number of unique values is less than or equal to 3.}
}
\value{
A data frame of the table.
}
\description{
Give a dataset and a list of variables, or just the data
in the vars.  For best results, convert categorical
variables into factors.
}
\examples{
# make me!!!
}
\keyword{utils}

