% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styler.R
\name{formatMedIQR}
\alias{formatMedIQR}
\title{Function to format the median and IQR of a variable}
\usage{
formatMedIQR(x, d = 2, na.rm = TRUE)
}
\arguments{
\item{x}{the data to have the median and IQR calculated}

\item{d}{How many digits to display. Defaults to 2.}

\item{na.rm}{Logical whether to remove missing values. Defaults to \code{TRUE}.}
}
\value{
A character string with results
}
\description{
Function to format the median and IQR of a variable
}
\examples{
formatMedIQR(mtcars$mpg)
}
\keyword{misc}
