% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{gglikert}
\alias{gglikert}
\title{Creates a plot for likert scale}
\usage{
gglikert(x, y, leftLab, rightLab, data, xlim, title, shape = 18, size = 7,
  colour = "grey50")
}
\arguments{
\item{x}{Variable to plot on the x axis (the likert scale responses or averages)}

\item{y}{The variable containing an index of the different items, should be integers}

\item{leftLab}{The variable with anchors for the low end of the Likert scale}

\item{rightLab}{The variable with anchors for the high end of the Likert scale}

\item{data}{The data to use for plotting}

\item{xlim}{A vector giving the lower an upper limit for the x axis.  This should be the
possible range of the Likert scale, not the actual range.}

\item{title}{A character vector giving the title for the plot}

\item{shape}{A number indicating the point shape, passed to \code{\link{geom_point}}}

\item{size}{A number indicating the size of points, passed to \code{\link{geom_point}}}

\item{colour}{A character string giving the colour of points, passed to \code{\link{geom_point}}}
}
\description{
Creates a plot for likert scale
}
\examples{

testdat <- data.frame(
  Var = 1:4,
  Mean = c(1.5, 3, 2.2, 4.6),
  Low = c("Happy", "Peaceful", "Excited", "Content"),
  High = c("Sad", "Angry", "Hopeless", "Anxious"),
  stringsAsFactors = FALSE)

gglikert("Mean", "Var", "Low", "High", testdat, xlim = c(1, 5),
  "Example Plot of Average Affect Ratings")

## clean up
rm(testdat)
}
