% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{internalformulaIt}
\alias{internalformulaIt}
\title{Internal function to create a formula}
\usage{
internalformulaIt(dv, iv, covariates)
}
\arguments{
\item{dv}{A character string of the dependent variable.}

\item{iv}{A character string or vector of the independent variables}

\item{covariates}{A character string or vector of the dependent variables}
}
\value{
A character string
}
\description{
This function is not intended to be called by users.
It creates a formula style character string from its argument.
But note that it does not actually create a formula class object.
If you do not want an argument, use the empty string.
}
\examples{
JWileymisc:::internalformulaIt("mpg", "hp", "am")
JWileymisc:::internalformulaIt("mpg", "hp", "")
JWileymisc:::internalformulaIt("mpg", "", "am")
}
\keyword{internal}
