% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{internalrunIt}
\alias{internalrunIt}
\title{Internal function to run a model using gam()}
\usage{
internalrunIt(formula, type, data, ...)
}
\arguments{
\item{formula}{A character string containing a formula style object.}

\item{type}{A character string indicating the type of dependent variable.
Currently \dQuote{normal}, \dQuote{binary}, or \dQuote{count}.}

\item{data}{A data frame to be used for analysis.}

\item{\dots}{Additional arguments passed to \code{gam}.}
}
\value{
A summary of the gam model.
}
\description{
This function is not intended to be called by users.
}
\keyword{internal}
