% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{intSigRegGraph}
\alias{intSigRegGraph}
\title{Function to find significant regions from an interaction}
\usage{
intSigRegGraph(object, predList, contrastList, xvar, varyvar,
  varyvar.levels, xlab = xvar, ylab = "Predicted Values", ratio = 1,
  xlim, ylim, xbreaks, xlabels = xbreaks, scale.x = c(m = 0, s = 1),
  scale.y = c(m = 0, s = 1), starts = 50)
}
\arguments{
\item{object}{A fitted rms object}

\item{predList}{TODO}

\item{contrastList}{TODO}

\item{xvar}{TODO}

\item{varyvar}{TODO}

\item{varyvar.levels}{TODO}

\item{xlab}{optional}

\item{ylab}{TODO}

\item{ratio}{TODO}

\item{xlim}{TODO}

\item{ylim}{TODO}

\item{xbreaks}{TODO}

\item{xlabels}{optional}

\item{scale.x}{optional}

\item{scale.y}{optional}

\item{starts}{Number of starting values to try between the
lower and upper bounds.}
}
\value{
A data table with notes if no convergence or significance
  thresholds (if any).
}
\description{
This function uses the \code{contrast} function from \pkg{rms} to
find the threshold for significance from interactions.
}
\examples{
## make me
}
