% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{is.naz}
\alias{is.naz}
\title{Is a variable missing, non finite or zero length character?}
\usage{
is.naz(x)
}
\arguments{
\item{x}{A vector to identify missing / non finite or zero length strings from}
}
\value{
a logical vector
}
\description{
Given a vector, return \code{TRUE} or \code{FALSE}
if each element is either missing (NA/NaN), non finite (e.g. infinite)
or a zero length character string (only for character vectors).
}
\examples{
is.naz(c(1, NA, NaN))
is.naz(c(1, NA, NaN, Inf))
is.naz(c("test", "", NA_character_))
}
\keyword{utils}
