% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{lagk}
\alias{lagk}
\title{Create a lagged variable}
\usage{
lagk(x, k = 1, by)
}
\arguments{
\item{x}{the variable to lag}

\item{k}{the length to lag it}

\item{by}{a variable to lag by. Must be sorted.}
}
\value{
a vector of the lagged values
}
\description{
Given a variable, create a k lagged version,
optionally do it by a grouping factor, such as an ID.
}
\examples{
lagk(1:4, 1)
}
\keyword{utils}
