% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{meanCircular}
\alias{meanCircular}
\title{Calculate a Circular Mean}
\usage{
meanCircular(x, max, na.rm = TRUE)
}
\arguments{
\item{x}{Numeric or integer values}

\item{max}{The theoretical maximum (e.g., if degrees, 360)}

\item{na.rm}{A logical value indicating whether to remove missing values.
Defaults to \code{TRUE}.}
}
\value{
A numeric value with the circular mean.
}
\description{
Function to calculate circular mean
}
\examples{
meanCircular(c(22:23, 1:2), max = 24)
meanCircular(c(12, 24), max = 24)
meanCircular(c(6, 7, 23), max = 24)
meanCircular(c(6, 7, 21), max = 24)
meanCircular(c(6, 21), max = 24)
meanCircular(c(6, 23), max = 24)
meanCircular(c(.91, .96, .05, .16), max = 1)

}
