% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{modelTest}
\alias{modelTest}
\alias{is.modelTest}
\alias{as.modelTest}
\alias{modelTest.vglm}
\alias{modelTest.lm}
\title{Detailed Tests on Models}
\usage{
modelTest(object, ...)

is.modelTest(x)

as.modelTest(x)

\method{modelTest}{vglm}(object, OR = TRUE, digits = 2L,
  pdigits = 3L, ...)

\method{modelTest}{lm}(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments passed to specific methods.}

\item{x}{A object (e.g., list or a modelTest object) to
test or attempt coercing to a modelTest object.}

\item{OR}{a logical value whether to report odds ratios and
95 percent confidence intervals, if \code{TRUE}, or
regression coefficients on the logit scale with standard
errors, if \code{FALSE}.}

\item{digits}{An integer indicating the number of digits for coefficients,
standard errors, and confidence intervals}

\item{pdigits}{An integer indicating the number of digits for
p-values.}
}
\value{
Depends on the method dispatch.

A list with two elements.
  \code{Results} contains a data table of the actual estimates.
  \code{Table} contains a nicely formatted character matrix.

A list with two elements.
  \code{Results} contains a data table of the actual estimates.
  \code{Table} contains a nicely formatted character matrix.
}
\description{
TODO: make me!
}
\examples{
mtcars$cyl <- factor(mtcars$cyl)
m <- VGAM::vglm(cyl ~ qsec,
  family = VGAM::multinomial(), data = mtcars)
modelTest(m)

rm(m, mtcars)

\dontrun{
mtcars$cyl <- factor(mtcars$cyl)
mtcars$am <- factor(mtcars$am)
m <- VGAM::vglm(cyl ~ qsec,
  family = VGAM::multinomial(), data = mtcars)
modelTest(m)

modelTest(m, digits = 4)$Table
modelTest(m, OR = FALSE)
modelTest(m, digits = 4, OR = FALSE)$Table

m <- VGAM::vglm(cyl ~ scale(qsec),
  family = VGAM::multinomial(), data = mtcars)
modelTest(m)

m2 <- VGAM::vglm(cyl ~ factor(vs) * scale(qsec),
  family = VGAM::multinomial(), data = mtcars)
modelTest(m2)

m <- VGAM::vglm(Species ~ Sepal.Length,
  family = VGAM::multinomial(), data = iris)
class(m)
modelTest(m)
}
m1 <- lm(mpg ~ qsec * hp, data = mtcars)
modelTest(m1)

mtcars$cyl <- factor(mtcars$cyl)
m2 <- lm(mpg ~ cyl, data = mtcars)
modelTest(m2)

## cleanup
rm(m1, m2, mtcars)
}
