% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{naz.omit}
\alias{naz.omit}
\title{Missing and Zero Character Omit}
\usage{
naz.omit(x)
}
\arguments{
\item{x}{A vector to exclude missing, non finite or zero length strings from}
}
\value{
a vector with missing/non finite/zero length strings omitted
}
\description{
Given a vector, exclude any missing values,
not a number values, non finite values,
and if a character class, any zero length strings.
}
\examples{
## stats na.omit
stats::na.omit(c(1, NA, NaN))
stats::na.omit(c("test", "", NA_character_))

naz.omit(c(1, NA, NaN))
naz.omit(c(1L, NA))
naz.omit(c(1L, NA, Inf))
naz.omit(c("test", "", NA_character_))
}
\keyword{utils}
