% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.residualDiagnostics}
\alias{plot.residualDiagnostics}
\title{Plot Residual Diagnostics Default Method}
\usage{
\method{plot}{residualDiagnostics}(x, y, plot = TRUE, ask = TRUE, ncol,
  ...)
}
\arguments{
\item{x}{A \code{residualDiagnostics} class object.}

\item{y}{Included to match the generic. Not used.}

\item{plot}{A logical value whether or not to plot the results or
simply return the graphical objects.}

\item{ask}{A logical whether to ask before changing plots.
Only applies to interactive environments.}

\item{ncol}{The number of columns to use for plots.
Missing by default which means individual plots are created.
If specified, plots are put together in a grid.}

\item{...}{Included to match the generic. Not used.}
}
\value{
a list including plots of the residuals,
  residuals versus fitted values
}
\description{
This function creates a number of diagnostic plots
from residuals. It is a default method.
}
\examples{
testm <- stats::lm(mpg ~ hp * factor(cyl), data = mtcars)

md <- residualDiagnostics(testm, ev.perc = .1)

plot(md)
plot(md, ncol = 2)

## clean up
rm(testm, md)
}
\keyword{plot}
