% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.SEMSummary.list}
\alias{plot.SEMSummary.list}
\title{Plots SEMSummary.list object}
\usage{
\method{plot}{SEMSummary.list}(x, y, which, plot = TRUE, ...)
}
\arguments{
\item{x}{An object of class SEMSummary.list.}

\item{y}{Ignored}

\item{which}{either a numeric vector based on the positions,
or a character vector giving the names of the levels of the
list to plot.}

\item{plot}{A logical, whether to actually plot the results or not.
Defaults to \code{TRUE}.}

\item{\dots}{Additional arguments passed on to the real workhorse, \code{corplot}.}
}
\description{
Plots SEMSummary.list object
}
\seealso{
\code{\link{corplot}}, \code{\link{SEMSummary}}
}
