% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{plot.SEMSummary}
\alias{plot.SEMSummary}
\title{Plots SEMSummary object}
\usage{
\method{plot}{SEMSummary}(x, y, ...)
}
\arguments{
\item{x}{An object of class SEMSummary.}

\item{y}{Ignored}

\item{\dots}{Additional arguments passed on to the real workhorse, \code{corplot}.}
}
\description{
Plots SEMSummary object
}
\examples{
# default plot
plot(SEMSummary(~ ., data = mtcars))

# same as default
plot(SEMSummary(~ ., data = mtcars), plot = "coverage")

# shows p values
plot(SEMSummary(~ ., data = mtcars), plot = "p")

# shows correlations
plot(SEMSummary(~ ., data = mtcars), plot = "cor")
}
\seealso{
\code{\link{corplot}}, \code{\link{SEMSummary}}
}
