% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plotting.R
\name{.plotDiagnosticsResiduals}
\alias{.plotDiagnosticsResiduals}
\title{Plot Residual Diagnostics}
\usage{
.plotDiagnosticsResiduals(object, ev.perc = 0.001)
}
\arguments{
\item{object}{A fitted model object from lm.}

\item{ev.perc}{A real number between 0 and 1 indicating the
proportion of the theoretical distribution beyond which
values are considered extreme values (possible outliers).
Defaults to .001.}
}
\value{
a list including plots of the residuals,
  residuals versus fitted values, and one list for
  plots of all random effects and finally a data table with
  any extreme values identified
}
\description{
This is an internal worker function and is not meant to be
called directly.
}
\examples{
# make me!
}
\keyword{internal}
\keyword{plot}
