% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/models.R
\name{R2}
\alias{R2}
\alias{R2.lm}
\title{Calculate R2 Values}
\usage{
R2(object, ...)

\method{R2}{lm}(object, ...)
}
\arguments{
\item{object}{A fitted model object.}

\item{...}{Additional arguments passed to specific methods.}
}
\value{
Depends on the method dispatch.

The raw and adjusted r-squared value.
}
\description{
Generic function to return variance explained (R2)
estimates from various models. In some cases these will be true
R2 values, in other cases they may be pseudo-R2 values if
R2 is not strictly defined for a model.
}
\examples{
R2(lm(mpg ~ qsec * hp, data = mtcars))
}
