% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{R2LMER}
\alias{R2LMER}
\title{Calculates the Marginal and Conditional R2 from lmer models}
\usage{
R2LMER(model, modelsum)
}
\arguments{
\item{model}{A model estimated by \code{lmer}.}

\item{modelsum}{The saved model summary (i.e., \code{summary(model)}).}
}
\value{
a named vector with the marginal and conditional R2 values
}
\description{
Calculates the Marginal and Conditional R2 from lmer models
}
\examples{
# make me!
}
\references{
For estimating the marginal and conditional R-squared values,
  see: Nakagawa, S. and Schielzeth, H. (2013). A general and simple method
  for obtaining R2 from generalized linear mixed-effects models.
  Methods in Ecology and Evolution, 4(2), 133-142. as well as:
  Johnson, P. C. (2014). Extension of Nakagawa & Schielzeth's R2GLMM to
  random slopes models. Methods in Ecology and Evolution, 5(9), 944-946.
}
\keyword{utils}
