% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{roundedfivenum}
\alias{roundedfivenum}
\title{Calculate a rounded five number summary}
\usage{
roundedfivenum(x, round = 2, sig = 3)
}
\arguments{
\item{x}{The data to have the summary calculated on}

\item{round}{The number of digits to try rounding}

\item{sig}{The number of significant digits to try}
}
\value{
The rounded or significant digit five number summary
}
\description{
Numbers are the minimum, 25th percentile, median,
75th percentile, and maximum, of the non missing data.
Values returned are either the significant digits or rounded values,
whichever ends up resulting in the fewest total digits.
}
\examples{
JWileymisc:::roundedfivenum(rnorm(1000))
JWileymisc:::roundedfivenum(mtcars$hp)
}
