% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives_helpers.R
\name{styledescriptives}
\alias{styledescriptives}
\title{Several internal functions to style descriptive statistics}
\arguments{
\item{n}{A character string giving the variable name (just the name)}

\item{x}{A variable (actual data, not just the name)}

\item{digits}{An integer indicating the number of significant digits to use.
Defaults to \code{2}.}

\item{includeLabel}{A logical value, whether or not to include the type of
descriptive statistics in the label/name.
Only applies to some functions. Defaults to \code{FALSE}.}
}
\value{
A data table with results.
}
\description{
Several internal functions to style descriptive statistics
}
\keyword{internal}
