% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/geoms_themes.R
\name{theme_tufte}
\alias{theme_tufte}
\title{Tufte Inspired Theme}
\usage{
theme_tufte(base_size = 11, base_family = "sans")
}
\arguments{
\item{base_size}{base font size, given in pts.}

\item{base_family}{base font family}
}
\description{
This is a barebones theme. It turns many aspects of plot background lines, etc. off completely.
Inspired from the excellent ggthemes package: https://github.com/jrnold/ggthemes
}
