% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/utils.R
\name{updateInstall}
\alias{updateInstall}
\title{Update R and install my core package set}
\usage{
updateInstall(x, repo)
}
\arguments{
\item{x}{A character vector of any additional packages to be installed}

\item{repo}{The repository to be used. Defaults to getOption("repos")}
}
\value{
NULL, called for its side effect.
}
\description{
Update R and install my core package set
}
\examples{
# updateInstall()
}
\keyword{utils}

