% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styler.R
\name{APAStyler.lm}
\alias{APAStyler.lm}
\title{APAStyler method for linear models}
\usage{
\method{APAStyler}{lm}(object, digits = 2, pdigits, file, ...)
}
\arguments{
\item{object}{A \code{lm} object}

\item{digits}{The number of digits to round results to. Defaults to 2.}

\item{pdigits}{The number of digits to use for p values. Defaults to digits + 1 if missing.}

\item{file}{An optional argument indicating whether the output should be written to a file.}

\item{...}{Additional argiuments passed on to \code{write.table}.}
}
\description{
APAStyler method for linear models
}
