% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/descriptives.R
\name{f.r2}
\alias{f.r2}
\title{Calculate F and p-value from the R2}
\usage{
f.r2(r2, numdf, dendf)
}
\arguments{
\item{r2}{r squareds}

\item{numdf}{numerator degrees of freedom}

\item{dendf}{denominator degrees of freedom}
}
\value{
a vector
}
\description{
Calculate F and p-value from the R2
}
\examples{
# make me!
}
\keyword{internal}
