% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/styler.R
\name{formatLMER}
\alias{formatLMER}
\title{Format results from a linear mixed model}
\usage{
formatLMER(list, modelnames, dig = 2)
}
\arguments{
\item{list}{A list of one (or more) models estimated from lmer}

\item{modelnames}{An (optional) vector of names to use in
the column headings for each model.}

\item{dig}{A numeric value indicating the number of digits to print.
This is still in early implementation stages and currently does not
change all parts of the output (which default to 2 decimals per
APA style).}
}
\value{
a data table of character data
}
\description{
Format results from a linear mixed model
}
\examples{
# make me!
}
\keyword{misc}
